Chargement des tiles Gestion du clavier Ajout de JPanel (c'est moche...) Refonte des collisions (Fonctionnelle mais immonde !!)
Todo
Gestion Clavier fluide (keypress->keyrealse avec une speed qui augmente a chaque "ticks")
Tiles + jolie
Et pleins de chose

Le C c'est mieux :)

+++ src/AnimatedThings.java
@@ -2,11 +2,11 @@ import java.awt.*;
 
 public class AnimatedThings extends SolidThings {
 
-    public AnimatedThings(int height, int width, int x, int y) {
+    public AnimatedThings(int height, int width, double x, double y) {
         super(height, width, x, y);
     }
 
-    public AnimatedThings(int x, int y, Image image) {
+    public AnimatedThings(double x, double y, Image image) {
         super(x, y, image);
     }
 }
